+++ CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
   - Update
     - Update `pubnub` dependency to `4.0.16`
     - Update gemspec to include dependency versions
+  - Bugfixes
+    - Fix `logger.level` default setting
 - **2016-09-05**: 1.3.4
   - Bugfixes
     - Fix `RingCentralSdk::REST::Subscription` renewal bug